Implementing ReadAllEventsForwardFilteredAsync
Implemented on both the server and client. Can filter by both stream and event type using either a prefix or a regular expression. Changed how the end of stream property on an event slice works to cope with the fact that after filtering has taken place there might not be any events in a given slice.

`maxSearchWindow` defines how many events should be searched before attempting to return the slice whether the required number of events are found or not.